# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Listing {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  address: string;
  city: string;
  state: string;
  zip: string;
  bedrooms: number;
  bathrooms: number;
  sqft: number;
  propertyType: string;
  status: string;
  images: string[];
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  parkingSpots?: number | null;
  features?: string[];
  isFeatured?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}

export interface ListingInput {
  title: string;
  description?: string;
  price: number;
  address: string;
  city: string;
  state: string;
  zip: string;
  bedrooms: number;
  bathrooms: number;
  sqft: number;
  propertyType: string;
  status: string;
  images?: string[];
  yearBuilt?: number;
  parkingSpots?: number;
  features?: string[];
  isFeatured?: boolean;
}

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  bedrooms?: number;
  bathrooms?: number;
  sqft?: number;
  propertyType?: string;
  status?: string;
  images?: string[];
  yearBuilt?: number;
  parkingSpots?: number;
  features?: string[];
  isFeatured?: boolean;
}

export interface PropertyTypeCount {
  propertyType: string;
  count: number;
}

export interface ListingStats {
  totalListings: number;
  avgPrice: number;
  forSaleCount: number;
  forRentCount: number;
  soldCount: number;
  byPropertyType: PropertyTypeCount[];
}

export interface SavedListing {
  id: number;
  listingId: number;
  createdAt: string;
  listing: Listing;
}

export interface SavedListingInput {
  listingId: number;
}

export type GetListingsParams = {
search?: string;
minPrice?: number;
maxPrice?: number;
minBedrooms?: number;
propertyType?: string;
status?: string;
city?: string;
};

export type GetRecentListingsParams = {
limit?: number;
};