# lib/api-zod/src/generated/types/listing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface Listing {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  address: string;
  city: string;
  state: string;
  zip: string;
  bedrooms: number;
  bathrooms: number;
  sqft: number;
  propertyType: string;
  status: string;
  images: string[];
  /** @nullable */
  yearBuilt?: number | null;
  /** @nullable */
  parkingSpots?: number | null;
  features?: string[];
  isFeatured?: boolean;
  createdAt: string;
  /** @nullable */
  updatedAt?: string | null;
}