# lib/api-zod/src/generated/types/listingStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PropertyTypeCount } from './propertyTypeCount';

export interface ListingStats {
  totalListings: number;
  avgPrice: number;
  forSaleCount: number;
  forRentCount: number;
  soldCount: number;
  byPropertyType: PropertyTypeCount[];
}