# lib/api-zod/src/generated/types/listingUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface ListingUpdate {
  title?: string;
  description?: string;
  price?: number;
  address?: string;
  city?: string;
  state?: string;
  zip?: string;
  bedrooms?: number;
  bathrooms?: number;
  sqft?: number;
  propertyType?: string;
  status?: string;
  images?: string[];
  yearBuilt?: number;
  parkingSpots?: number;
  features?: string[];
  isFeatured?: boolean;
}