# lib/api-zod/src/generated/types/savedListing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Listing } from './listing';

export interface SavedListing {
  id: number;
  listingId: number;
  createdAt: string;
  listing: Listing;
}