# lib/api-zod/src/generated/types/savedListingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Real Estate Listing Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SavedListingInput {
  listingId: number;
}